feature(kafka): 	1. 카프카 연결
issue: #210

+++ backend/src/main/java/com/patriot/fourlipsclover/match/dto/request/LocalsProposalRequest.java
@@ -14,7 +14,7 @@ import java.util.List;
 public class LocalsProposalRequest {
 
     private Integer matchId;
-    private List<Long> restaurantIds;
+    private List<Integer> restaurantIds;
     private String recommendMenu;
     private String description;
