+++ backend/src/main/java/com/patriot/fourlipsclover/restaurant/repository/RestaurantJpaRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface RestaurantJpaRepository extends JpaRepository<Restaurant, Long> {
+public interface RestaurantJpaRepository extends JpaRepository<Restaurant, Integer> {
 
     Optional<Restaurant> findByKakaoPlaceId(String kakaoPlaceId);
